Add files via upload
Report of recorded values for watchlist members, open, close, high and low prices are outputted and the Python API returns the mean as "analysis" value in Json.  The report format is shown in this one while the python file is attached separately.  The python file shows the use of list to collect the data, the len function and the sum functions combined for the mean calc.