Add files via upload
Understanding Datawarehouses often involves such things as queries in sql to delimit the information required and such things as what am I trying to visualize better.  In some cases what we want to examine are the reasonable ranges of expected values or what appear to be ranges.  For example what are normal ranges of temperature in October in new England?  We would want to understand the numbers and what amounts to a good data set.  Data sets are often stored in separate database repositories, workbooks etc.  Let's look at an example where we import data from two repositories, what is involved and why the benefit occurs?